Remove code supporting Gingerbread and below.
