+++ app/src/main/java/com/totsp/crossword/net/DownloadReceiver.java
@@ -1,13 +1,13 @@
 package com.totsp.crossword.net;
 
-import java.util.HashMap;
-
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
 
 import com.totsp.crossword.puz.PuzzleMeta;
+
+import java.util.HashMap;
 
 public class DownloadReceiver extends BroadcastReceiver {
 	
@@ -16,13 +16,11 @@ public class DownloadReceiver extends BroadcastReceiver {
 	
 	private BroadcastReceiver impl; 
 	{
-		if(android.os.Build.VERSION.SDK_INT >= 9){
-			try{
-				BroadcastReceiver built = (BroadcastReceiver) Class.forName("com.totsp.crossword.net.DownloadReceiverGinger").newInstance();
-				impl = built;
-			} catch(Exception e){
-				e.printStackTrace();
-			}
+		try{
+			BroadcastReceiver built = (BroadcastReceiver) Class.forName("com.totsp.crossword.net.DownloadReceiverGinger").newInstance();
+			impl = built;
+		} catch(Exception e){
+			e.printStackTrace();
 		}
 		if(impl == null){
 			impl = new DownloadReceiverNoop();
