+++ app/src/main/java/com/totsp/crossword/view/recycler/ScrollDetector.java
@@ -1,11 +1,10 @@
 package com.totsp.crossword.view.recycler;
 
         import android.content.Context;
-        import android.os.Build;
-        import android.view.GestureDetector;
-        import android.view.MotionEvent;
-        import android.view.View;
-        import android.view.ViewConfiguration;
+import android.view.GestureDetector;
+import android.view.MotionEvent;
+import android.view.View;
+import android.view.ViewConfiguration;
 
 public abstract class ScrollDetector extends GestureDetector.SimpleOnGestureListener implements View.OnTouchListener {
     private final GestureDetector mDetector;
@@ -29,12 +28,7 @@ public abstract class ScrollDetector extends GestureDetector.SimpleOnGestureList
     }
 
     protected int getSlop(Context context) {
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.FROYO) {
-            //noinspection deprecation
-            return ViewConfiguration.getTouchSlop() * 2;
-        } else {
-            return ViewConfiguration.get(context).getScaledPagingTouchSlop();
-        }
+        return ViewConfiguration.get(context).getScaledPagingTouchSlop();
     }
 
     @Override
